requirements.txt
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
filelock==3.18.0
fire==0.7.0
fsspec==2025.7.0
hf-xet==1.1.5
huggingface-hub==0.33.4
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
numpy==1.26.4
opencv-python==4.9.0.80
packaging==25.0
pillow==11.3.0
pybboxes==0.1.6
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
sahi==0.11.31
shapely==2.1.1
sympy==1.13.1
termcolor==3.1.0
terminaltables==3.1.10
tokenizers==0.21.2
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
transformers==4.53.2
typing_extensions==4.14.1
urllib3==2.5.0
